[6CORE-101] moved again within this task
