+++ 3-platform/warehouse/infra/src/main/java/com/_6core/platform/warehouseinfra/mapper/ImmutableWarehouse2WarehouseEntity.java
@@ -2,7 +2,7 @@ package com._6core.platform.warehouseinfra.mapper;
 
 import com._6core.lib.java.domain.model.warehouse.WarehouseV01;
 import com._6core.lib.java.domain.model.warehouse.immutable.ImmutableWarehouseV01Impl;
-import com._6core.platform.warehouseinfra.adapter.driven.persistent.entity.WarehouseEntity;
+import com._6core.platform.warehouseinfra.adapter.driven.driven.persistence.entity.WarehouseEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
